+++ registry/storage/driver/middleware/googlecdn/middleware.go
@@ -36,16 +36,16 @@ const defaultDuration = 20 * time.Minute
 
 // customGitlabGoogle... are the query params appended to googlecdn signed redirect url
 const (
-	customGitlabGoogleNamespaceParam  = "gitlab-namespace"
-	customGitlabGoogleProjectParam    = "gitlab-project"
-	customGitlabGoogleAuthTypeParam   = "gitlab-auth-type"
-	customGitlabGoogleObjectSizeParam = "gitlab-size-bytes"
+	customGitlabGoogleNamespaceIdParam = "gitlab-namespace-id"
+	customGitlabGoogleProjectIdParam   = "gitlab-project-id"
+	customGitlabGoogleAuthTypeParam    = "gitlab-auth-type"
+	customGitlabGoogleObjectSizeParam  = "gitlab-size-bytes"
 )
 
-// customParamKeys is the mappping between gitlab keys to googlecdn signed-redirect-url query parameter keys
+// customParamKeys is the mapping between gitlab keys to googlecdn signed-redirect-url query parameter keys
 var customParamKeys = map[string]string{
-	dstorage.NamespaceKey:   customGitlabGoogleNamespaceParam,
-	dstorage.ProjectPathKey: customGitlabGoogleProjectParam,
+	dstorage.NamespaceIdKey: customGitlabGoogleNamespaceIdParam,
+	dstorage.ProjectIdKey:   customGitlabGoogleProjectIdParam,
 	dstorage.AuthTypeKey:    customGitlabGoogleAuthTypeParam,
 	dstorage.SizeBytesKey:   customGitlabGoogleObjectSizeParam,
 }