+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## [4.0.0](https://gitlab.com/gitlab-org/container-registry/compare/v3.93.0-gitlab...v4.0.0-gitlab) (2024-4-29)
+
+
+### ⚠ BREAKING CHANGES
+
+* **storage:** remove OSS and Swift storage drivers
+* **v2:** remove old tag delete api
+
+### ✨ Features ✨
+
+* **cache:** redis repository cache for list repository tags operation ([9ea6ebe](https://gitlab.com/gitlab-org/container-registry/commit/9ea6ebe2b4f6f0a7fdb85da6787e8d9b79bd6107))
+* **storage:** remove OSS and Swift storage drivers ([32f4118](https://gitlab.com/gitlab-org/container-registry/commit/32f4118056300b9afaa755dd1faf35b9c9d356d2))
+* **v2:** remove old tag delete api ([df7dbb0](https://gitlab.com/gitlab-org/container-registry/commit/df7dbb01d33ed705b0745153e42a6e8388e86112))
+
+## [3.93.0](https://gitlab.com/gitlab-org/container-registry/compare/v3.92.0-gitlab...v3.93.0-gitlab) (2024-04-23)
+
+
+### ✨ Features ✨
+
+* inject GitLab namespace and project IDs into CGS/CDN signed URLs ([ec9c72e](https://gitlab.com/gitlab-org/container-registry/commit/ec9c72e74bf8a1a78490c46abfe3f8f7188ac65d))
+* **notifications:** add backoff sink with maxretries ([fa1a985](https://gitlab.com/gitlab-org/container-registry/commit/fa1a985737fcfe18b64943a1018c3535fcfc448e))
+* **s3wrapper:** retry delete s3 objects operation if incomplete ([d911653](https://gitlab.com/gitlab-org/container-registry/commit/d91165374e8759bd0d05176ad705c3b8515c975a))
+
 ## [3.92.0](https://gitlab.com/gitlab-org/container-registry/compare/v3.91.0-gitlab...v3.92.0-gitlab) (2024-04-04)